restructure -> move Backbone to top level backbone package
