+++ core/src/test/scala/backbone/BackboneSpec.scala
@@ -1,14 +1,11 @@
 package backbone
 
+import backbone.Backbone.Consumed
 import backbone.format.Format
-import backbone.scaladsl.{Backbone, ConsumerSettings}
-import backbone.scaladsl.Backbone.Consumed
 import com.amazonaws.services.sns.AmazonSNSAsyncClient
 import org.scalatest.WordSpec
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.mockito.MockitoSugar
-
-import scala.util.{Success, Try}
 class BackboneSpec extends WordSpec with DefaultTestContext with MockitoSugar with ScalaFutures {
 
   private[this] implicit val snsClient = mock[AmazonSNSAsyncClient]
